유저 채팅 토큰 검증 로직 구현완료
