--- src/main/java/newCar/socket_app/model/NoticeMessage.java
@@ -1,4 +1,0 @@
-package newCar.socket_app.model;
-
-public class NoticeMessage {
-}
