+++ src/main/java/newCar/socket_app/model/chat/BlockMessage.java
@@ -1,4 +1,4 @@
-package newCar.socket_app.model;
+package newCar.socket_app.model.chat;
 
 public class BlockMessage {
     private String id;
