+++ src/main/java/newCar/socket_app/model/chat/NoticeMessage.java
@@ -1,0 +1,15 @@
+package newCar.socket_app.model.chat;
+
+import com.fasterxml.jackson.annotation.JsonTypeInfo;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter @Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@JsonTypeInfo(use = JsonTypeInfo.Id.NONE)
+public class NoticeMessage {
+    private String content;
+}
