+++ src/main/java/newCar/socket_app/model/chat/UserChatMessage.java
@@ -1,0 +1,5 @@
+package newCar.socket_app.model.chat;
+
+public class UserChatMessage extends ChatMessage {
+
+}
