+++ src/main/java/newCar/socket_app/model/session/AdminSession.java
@@ -1,4 +1,5 @@
 package newCar.socket_app.model.session;
 
 public class AdminSession extends Session{
+
 }
