+++ src/main/java/newCar/socket_app/model/session/UserSession.java
@@ -1,8 +1,11 @@
 package newCar.socket_app.model.session;
 
 import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
 import newCar.socket_app.model.Team;
 
+@Getter @Setter
 @AllArgsConstructor
 public class UserSession extends Session {
     private Long userId;
