+++ src/main/java/newCar/socket_app/service/game/GameStateService.java
@@ -1,6 +1,6 @@
 package newCar.socket_app.service.game;
 
-import newCar.socket_app.model.GameData;
+import newCar.socket_app.model.game.GameData;
 
 public interface GameStateService {
     public void updateGameState(GameData gameData);
