+++ src/main/java/newCar/socket_app/service/game/RedisGameStateManager.java
@@ -3,8 +3,8 @@ package newCar.socket_app.service.game;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import newCar.socket_app.model.GameData;
-import newCar.socket_app.model.GameState;
+import newCar.socket_app.model.game.GameData;
+import newCar.socket_app.model.game.GameState;
 import newCar.socket_app.model.Team;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
