+++ src/main/java/newCar/socket_app/service/message/MessagePublisherService.java
@@ -1,6 +1,6 @@
 package newCar.socket_app.service.message;
 
-import newCar.socket_app.model.ChatMessage;
+import newCar.socket_app.model.chat.ChatMessage;
 
 public interface MessagePublisherService {
     public void publish(String topic, ChatMessage message);
